Implemented TerrainManager class to handle multiple terrains. This allows us to get the height of the terrain no matter where the player may be in the world.
Currently we are using a grid terrain matrix row x col.
TODO: In the future, I think it might be better to switch to a hex based mapping terrain. This could be done with a new HexTerrain structure with six pointers to the adjoining terrains.

It might be difficult to determine which triangle the player is on using a hex based terrain system... need to think about this some.